requirements.txt
# Core ML
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnxruntime==1.16.0

# Data processing
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2
tqdm==4.66.1

# FastAPI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# TorchServe
torch-model-archiver==0.9.0
torchserve==0.9.0

# Testing
pytest==7.4.3
locust==2.19.1

# Config
pyyaml==6.0.1